change to ProductA (implementation behind factory) signature, no changes to any clients

+++ src/no_factory/ProductA.java
@@ -5,7 +5,7 @@ import simple_factory.IProduct;
 public class ProductA implements IProduct {
     private ProductA() {}
     /* funny constructor */
-    public ProductA(int i) {
+    public ProductA(float i) {
         System.out.println("i=" + i);
     }
 }
